Perf: Unify the styles of directly connected pages and pages that open in a new window.
